Use standard tag ring mechanism for jumping to and back from definitions
Slime (and my elisp-slime-nav library) use the built-in etags
mechanism for maintaining a poppable stack of jump locations. This
commits simplifies nrepl.el's jump mechanism by using the same trick.

Additionally, this commit solves a problem whereby a jump location
would unnecessarily be pushed when trying to jump to the non-existent
definition of a bogus or undefined symbol.