Fix #211 - implement jump to error from stacktrace
Now when someone gets a compilation error after `C-c C-k` the top error
of the stacktrace is clickable and it will take the user to the
offending line in the source code. This doesn't work when there is
no source file (say after eval last sexp) for obvious reasons. The functionality
relies on `compilation-minor-mode` being enabled in the error buffer (which is done
automatically after it's created).