Highlight compilation error in the source buffer
The compilation error's line number is extracted from the stacktrace
buffer and highlighted with an overlay in the related source buffer.
The error overlay is cleared upon recompilation. There is some race condition
between the moment in which we start to parse the stacktrace buffer and the moment
the stacktrace has been output to it, therefore the 0.3s sleep to suppress it.