Add multiple REPL connections
Allows REPLs for several different nREPL servers at once.

The default connection is used to dispatch nrepl commands in clojure buffers.
Use the nrepl-make-repl-connection-default command to make a specific repl
session the default.

Adds a nrepl-close command to close the current connection.

The current buffer is returned by the nrepl-current-connection-buffer function.
nrepl-connected-list maintains a list of connections.  When closing REPLS,
this is used to decide whether to disable nrepl interaction mode on clojure
buffers, and in order to set a new default connection if one is available.

Implements #75.