Improve compilation error handling
To avoid delays associated with retrieval of the stacktrace, highlighting is
based on the exception message that is reported by the stderr-handler in
nrepl-load-file-handler.  The exception trace is put in the *nrepl-error*
buffer, but not displayed by default.

Adds an nrepl-visit-error-buffer function, to visit *nrepl-error*.

Installs nrepl-jump-to-compilation-error as the next-error-function in
nrepl-interaction-mode buffers, to jump to the compilation error using
next-error.

Adds an optional eval-error-handler to nrepl-make-response-handler.

Adds nrepl-popup-on-error, to control whether stacktraces (globally) are popped
up by default.