Automatically make lein available.
nrepl-server-command is now modified to take either 'search
or a string. Access is now via nrepl-fetch-server-command. If,
nrepl-server-command is a string, use this. If search, look for the
nrepl executable. If it is not found an error message is signalled.
As nrepl-fetch-server-command returns full paths, launch through a
login shell is no longer necessary.

A new command, nrepl-lein-self-install, is provided which fetches the
lein script (shell or batch file as appropriate) and installs it in
user-emacs-directory.