Adding support for remote files that are loaded using Tramp, eg. over ssh/scp.
If nrepl is started from a buffer that is loaded with a remote file, the
tramp method used for loading the file is associated with the nrepl
connection such that file and archive (jar/zip) references are resolved
through the same tramp method. This helps improve remote development.

For remote archives, this can be slower than loading archives locally,
by mapping paths to a local Maven repository for instance. But this is a
trade-off made to protect the accuracy of the library data, which is
often critical when debugging a remote service.