add cider connection hooks on autoload of cider
We have connected/disconnected hooks for nrepl-client, but really they
are being used by cider itself. We can setup the hooks correctly when we
autoload cider.el, and all of the dependencies will have loaded. As a
result it's not necessary to autoload the connect/disconnect actions for
cider as by that point in time they will be loaded due to the manual
require in cider.el of cider-interaction.el.