Improve support for tramp source buffers
Ensures load-file doesn't use the tramp location of the file for the
filename passed to the nrepl load-file op (which gets compiled into
debug information).  This caused some issues when using the same repl
with source files both locally and over tramp.

Uses an explicit buffer to determine the tramp prefix rather than the
find-tag-marker-ring.  Changes the signature of cider-jump-to-def-for to
add a buffer argument.

Changes cider-jump-to-def-for to look for a tramp path before a local
path.