Use cider variants of lisp-eval-region and eval-last-sexp in repl
It's an odd use case, but using C-c C-r or C-x C-e inside of a
cider repl should not throw an error about missing an inferior lisp
process. It is useful for inspecting values in the middle of a longer
sexp. This adds the same bindings used in the cider minor mode to the repl.