Shortcut for cider-pprint-eval-last-sexp in the repl
Per discussion at #535, I tried using `cider-pprint-eval-defun-at-point` bound
to `C-c C-f` as it is in the interactive minor mode. Unfortunately
defun-at-point is not as well defined at the repl for some reason, I noticed
similar while experimenting with `cider-eval-defun-at-point` at the repl. So I
tried it with `cider-pprint-eval-last-sexp`, as that seemed more useful. I
guess the mnemonic would be see form?

This at least improves the case where the repl returns a very large vector, map
or quoted list and it needs to be inspected visually. It doesn't reformat it
in-place but it is reformatted in the popup buffer. I also think this plays well
with the exiting pprint toggle in the repl as a shortcut.