implement cider-eval-last-sexp-value-prefix
When using `cider-eval-last-sexp' it is often desireable to not
have the resulting value inserted as an expression but as a
comment. This commit implements two customizable variables,
`cider-interactive-eval-print-value-prefix' and `cider-eval-
last-sexp-eval-value-prefix', which allow the user to control
what strings prefix inserted values both in the general case of
inserting values into a buffer and the particular case of `cider-
eval-last-sexp' being called with a prefix argument.