Don't move point when clearing or highlighting test results
When running tests with cider-test failing tests are highlighted in the source
files, or if tests are passing again previous highlights are cleared. The code
that does this is using the `cider-find-var` function to find the buffer for a
vars under test and then does it's job.

`cider-find-var` used to move point to the line at which the var was found. This
is a problematic side effect, because code using `cider-find-var` to find the
buffer can't use save-excursion because it doesn't know which buffer to safe
beforehand.

This PR removes the `goto-line` in `cider-find-var` and returns just the buffer
without any side effect. The `cider-company-location` uses similar code as
`cider-find-var` but with `save-excursion` wrapped around it.