Clean symbol before doing look up
If you attempt to run `cider-jump-to-var` to find the definition of an
atom dereferenced with `@`, cider can't find the source location. If you remove
the `@` cider has no issues finding the location. This allows for cleaning
non-variable components out of a symbol before doing the look up to prevent this
issue.