Fix ansi output coloring for test failures
The function cider-repl--input-line-beginning-position used to return the
position at the beginning of the line containing cider-repl-input-start-mark.
Commit 83d360d made it impossible to move the cursor to this position IF the
marker was on the same line as the prompt. This caused text to be inserted
in the wrong position and not have its color codes processed. Since all callers
of this function subtract one from its result to find the end of the line that
precedes that of the marker, we simply make a function that does this directly
and avoids the prompt issue.

Fixes #968