Fix cider-insert-in-repl indenting too much
The intention was to indent the code that was just inserted.  The bug was that
the code grabbed the value of point from the source buffer instead of the repl
buffer to mark the start position when indenting.  The result, in most cases,
was that the entire repl sessions was re-indented.  As the repl can grow quite
long this often brought emacs to a grinding halt.