[Fix #1139] Correct the handling of ns forms and forms with unevaluted ns
Ns forms should always be evaluated in the "user" namespace. We have introduced
a regression in 0.9.0 after which they started being evaluated in their own
namespace, which doesn't exist originally. This commit also fixes a problem
with the auto-evaluation of ns forms, when a form from their ns gets evaluated
before the ns form.