cider-interactive-eval takes a BOUNDS argument
interactive-eval takes a bounds third argument (instead of a point).
This allows it to properly provide a good marker to
display-eval-result (isntead of just giving it the point-marker), so
the result of the evaluation is always displayed at the end of the
sexp, instead of at line where point happens to be located.