Improve usage of the eval result overlays
Change display-eval-result to be smarter when it receives a marker, so
it does the right thing when the original buffer was killed before
evaluation finished.

If the result overlay would be out of view (or in a hidden buffer) it
will also display the result as a message regardless of user option.

Hiding the message is better handled now. Instead of messaging the
result and then messaging nil to hide it, we simply message the result
with an invisible property. This means it's not displayed in the echo
area, but shows up on the Messages buffer.