[Fix #1252] Fix breakage in cider-repl-clear-buffer
After running cider-repl-clear-buffer and then loading a file creating
output (e.g. because it contains println statements) the output would
appear in the wrong place, causing the prompt to disappear.

The problem was in the function responsible for putting POINT in the right
place before emitting evaluation results into the repl buffer.  When the
prompt was the first line in the buffer,
cider-repl--end-of-line-before-input-start didn't move point at all.

cider-repl--end-of-line-before-input-start is now changed to move backward
past the cider-repl-prompt field in order to find the appropriate output
position.