Simplify connection tracking code
Might affect #1266 and #1263.

`cider-find-connection-buffer-for-project-directory` embraces
ambiguities instead of avoiding them. By returning the first of multiple
matching buffers, it plays nice with the sibling connections.
It also returns buffers with nil directories if nothing else matches,
but this is just to try to fail as little as possible, it has nothing to
do with siblings.

`cider-switch-to-relevant-repl-buffer` just calls
`cider-current-repl-buffer` directly. This allows it to work in sibling
connections without complaining.