[Fix #467] Improve completion in the repl
When investigating #467 I noticed that the test checking if we're in a
string or not had a mostly false positves in the repl.

The previous solution relied on the parse state from point-min up to
point which I think is too large a context.  Furthermore, it's not
reasonable to expect the entire contents of the repl buffer, prompts and
all, to be valid lisp.