Fix lag in the debugger
The lag was happening because we waited for process-output twice: once
for the :done message and once for the next input request. Sadly,
a lot of times these two messages arrived together (as far as Emacs
could tell), so the second wait just did nothing until it timed out.

cider-nrepl no longer sends the :done message, and so we only need to
wait once now.