Work on the assumption that cider-repl-type is never nil
This is partially to simplify code, and partially to make it easier for
us to catch bugs early.
Since the default value of cider-repl-type is "clj", the current
behaviour shouldn't change for users. But we can do
    (setq-default cider-repl-type nil)
as a way to ensure that any flaws in the state tracker will fail early.