Give better advice for *print-length* setting
The README.md suggested that it would be a good idea to globally set `*print-length*` through the `:global-vars` key.  I followed the advice and have been bitten by that several times, because a non-`nil` `*print-length*` setting can cause several libraries (for example alembic or anything clojurescript related) to be incorrectly compiled.  In theory, everyone should make sure that the printer related dynamic variables are in a sane state when relying on the printer output to be something specific, but in practice this is hard to do, also because there is no macro like CL's `with-standard-io-syntax` that binds all printer dynamic variables to a sane state.