Slightly reduce the padding on stacktrace file names
Was 30, now 26. The longest file name on all of my stacktraces is always
ThreadPoolExecutor.java, which has 23 chars and isn't even displayed
with the default filters. With the new setting, this file will still be
indented by 3 spaces.

The reason is that error buffers are usually created on a split window,
and stacktraces are indented by so many spaces that I can never read the
function names without unsplitting the window first.

Maybe we can reduce it even more.