Make local Javadoc lookup paths configurable
When opening the javadoc of a class, CIDER will try to browse to a
relative url of the shape `java/lang/String.html`.

Add a new customizable variable where lookup paths can be configured.
These will be checked one by one to find the right HTML file, or if none
found will fall back to the old behavior.