Let sync-call errors use the stacktrace buffer
Sync-call errors do not bind *e unlike eval-call errors. This made it
impossible to use the stacktrace middleware to create the nicely
analyzed and formatted stacktrace buffer (see Issue #1420
https://github.com/clojure-emacs/cider/issues/1420).

In order to rectify this, the nREPL code was modified slightly to store
sync-call related errors and provide a mechanism for the stacktrace
middleware to retrieve and process these stored exceptions. This results
in the ability for sync-call related errors to be analyzed and presented
using the standard stacktract buffer.