Add default value to nrepl-dict-get
Make the nrepl-dict hashes act similarly to the Clojure maps by giving
them an optional default return value. If the provided key is not found
in the dict, then the optional default value will be returned. Notably,
the nil value is allowed as both a key and a value in nrepl-dict's.

Updated `cider--check-middleware-compatibility` to use this
functionality.