Fix force-step operations in the debugger
Two operations, `:here` and `:out` currently support a `force?`
parameter, which causes the debugger to skip breakpoints in other
instrumented functions until reaching the specified point.

They are bound to uppercase letters, `H` and `O` respectively. Neither
was working previously, because the `force?` parameter was not being
sent (in the case of `:here`), or was being sent incorrectly (in the
case of `:out`).