[Fix #1561] Correct font-lock vars for namespaces not loaded in the REPL
The current ns-browser relies on the `cider-repl-ns-cache` to decide font-locks. This gives us incorrect font-locks for namespaces which are present on the classpath, but not loaded in the REPL.
We change this to get the data from the nREPL middleware op `ns-vars-with-meta`,in case of a cache miss; clojure-emacs/cider-nrepl#346.