Make cider-jack-in-clojurescript honor .dir-locals.el
It used to be a handy trick to set cider-cljs-lein/boot-repl in .dir-locals.el,
but this no longer works. This patch fixes that so this once again works.

The context when spawning the CLJS repl is the nrepl-server process buffer, and
since it does not visit a file it ignores .dir-locals.el. With
`hack-local-variables` we can explicitly load them.

Fixes #1938