[Fix #1947] Inhibit `enlighten-mode` during repl initialization
When initializing the repl, we ask what namespace we are in by evaling code,
eg (eval "(str *ns*)"). However, enlighten mode hijacks eval messages in the
middleware and uses the debugger which causes an error if the debugger is not
yet initialized, ie, in `cider-repl-init`.

So we inhibit this feature. The reason I am doing this let binding rather than
just reordering is that reordering is brittle and doesn't convey that this order
is very important. While its kinda gross that we are dynamically altering this
variable several call stacks up, this is a very special case right at startup
and not complicated logic during the principal use of CIDER.