Fix expectations in cider-font-lock-tests
These tests pass in their current form, but they do not pass on a build of emacs
from source tracking master. The tests use cl-every and cl-some and the tests
assert that the results are either `:equal-to t` or `:not :equal-to t`. Right
now, cl-some returns t but on my build of emacs from master it returns the truth
value that it finds, causing buttercup to fail in its comparison with:

FAILED: Expected cider-reader-conditional-face to be `eq' to t

So rather than checking against some value for truth, we use the buttercup
assertion :to-be-truthy.

Further, we remove the reliance on the cl library and switch over to seq.el.

Also some line length formatting and rename test predicate helpers.