Show only or hide when on stackframe lines
In order to add "positive" filters, we must only show or hide when we're
actually on a stackframe line. This method takes a very naive way of starting at
the top of the buffer and considering whether to hide all lines. The other lines
just don't have any 'flags properties so they never come up as needing to be
hidden in the `(seq-intersection filters flags)` part, so hide is always false
on them. But if we want to show _only_ those lines with a particular flag this
hides the cause line and any other lines that don't have flags at all.