[Fix #2148] Fix jump to definition for remote cider-jack-in or cider-connect (#2156)
The `cider-tramp-prefix` does not return only tramp prefix.
It used the `tramp-make-tramp-file` in the `with-parsed-tramp-file-name` which force it to make a full tramp file path, not only prefix. I placed the `cider-make-tramp-prefix` for making prefix only. It is modified a little bit from `tramp-make-tramp-file`.

The `cider--client-tramp-filename` does not remove `file: scheme` for tramp filename.
It should return the string with tramp prefix and remote source path. But previous version leaves `file: scheme` in a remote source path. I just added a line to remove this scheme in this function.

The `cider-find-file` leaves a useless list buffer per opening remote archived source like zip, jar.
When continuously tracking, it'll will generate a lot of trash and make us tired to remove it. Now it'll be removed automatically.