[Fix #1976] Add an interactive command for connecting to a running ClojureScript REPL
This worked in the past as well, but someone had to toggle the type of the REPL
to "cljs" using `cider-repl-type` which was not obvious to most people. I also
added an interactive command for toggling the REPL type - `cider-repl-set-type`.

This fixes a mess I had created before - I had copy-pasted some code I planned
to tweak from the related issue and I committed this code upstream by mistake.