[#2202] Check in advance if a ClojureScript REPL can be created or not
The current logic relies mostly on namespace checks, but it can
certainly be improved down the road.

The most important thing is that now users are going to get more
meaningful errors earlier. Before this we'd first create the second
REPL buffer and just get there some obscure errors because the command
didn't check at all if piggieback or the target REPL were even
available.