Provide a way to specify known locations for .nrepl-port lookup
If a boot process is spawned from a directory that doesn't house a boot project
the proccess will not indicate where it was spawned from.  This change adds an
additional list custom var `cider-nrepl-port-additional-search-paths` that the
user can configure in the cases they commonly spawn a repl server from a
particular directory and want to easily connect to the running repl server.  It
simply appends that path to the list of paths to paths it retrieves in
`cider-locate-running-nrepl-ports`.