Add test selector support
Test all and test ns understand leiningen style test selectors, eg
metadata markers on `deftest`s. No support for namespace level metada
markers for now.

If test selector is specified only the test(s) run that have that
selector in their metadata.

If all loaded tests or project tests are invoked with prefix, ask for
test selector in the minibuffer. A command is available to run ns
tests with a selector.