Adding cider-nrepl-port-search-paths defcustom
This defcustom is useful for cases when you want to connect to a local repl but
are using boot repl spawned from a location that is not a project folder.
The boot process does not display the start directory so we cannot search for
the nrepl-port file from the process. This defcustom allows you to list common
places where you would spawn your boot repl and add that to the search path when
looking for nrepl-port files to show in.

Fixing indentation error.