[Fix #2294] simplify stacktrace filter mechanism
previously there were two separate lists of positive and negative
filters and complicated logic spread throughout of how these
interacted. Now there is a single list that you can toggle flags on or
off and their interaction is dictated inside of a single function
`cider-stacktrace-apply-filters`. Since their interaction is here we
no longer need to remember previous flags as they are just toggled on
and off.

Positive filters are 'all or 'project. These dictate that all frames
should be shown or only any frame from a project namespace should be
shown, despite other filters. Also, these two cannot exist at the same
time, so we need to check which is most "recent". Since if 'project is
selected, it hides all but project frames and if 'all is selected it
must show all. We do what was most recently selected in this case
then.

So the frame filtering is quite simply a list of tags you want to
filter by: '(java clj tooling repl dup project all). These are toggled
on and off as desired.