Defcustom to allow inverting eval with prefix on insert
On current insert commands, default behavior is to insert the form and
not evaluate. A prefix arg will evaluate the form after it is placed
in the repl. This will allow to invert this behavior, defaulting to
evaluating the form when inserted and interpreting the prefix arg as
preventing evaluation.