Group insert-into-repl commands in a common keymap and make them more flexible  (#2319)
I noticed the insert-into-repl commands didn't have an easy map to use those commands so I added them under a map. Allows some defcustoms to tweak the behavior:

- `cider-invert-insert-eval-p`: Previously, the prefix arg was required to have the inserted form evaluated and the default behavior was to just plop the form in the repl. This inverts that so that by default the forms are inserted and then evaluated. The benefit is that now the repl has a sequence of values and the forms that give those values context.
- `cider-switch-to-repl-after-insert-p`: Previously when using an insert command point was moved into the repl buffer unconditionally. Now it is conditional. This is desirable since now that it is easy to have inserted forms evaluated by default the repl often doesn't need focus after each insertion.

The motivation behind these changes is when working with `comment` forms.

Imagine some comment forms that are useful for inspecting your namespace. This makes it easy to eval a sequence of them and put their forms and results in the repl so there is a sense of context. It is also easy to put a prefix arg on them to prevent evaluation and then modify them if you need.

This does some work towards @halgari 's issue in https://github.com/clojure-emacs/cider/issues/2255 in putting a similar behavior to Cursive into CIDER. Notice how the evaluations from the comment form are plopped into the repl to give a sense of context to the state of the repl. That workflow now will work with the caveat that you have to be at the end of the form rather than any point inside of it (but that is next on the roadmap). This felt like a way to get 70% of the functionality there quickly.

To emulate the behavior of that gif, do

```elisp
(setq cider-invert-insert-eval-p t)            ;; eval after inserting by default
(setq cider-switch-to-repl-after-insert-p nil) ;; don't switch to the repl buffer
```

Then the insert map is now at `C-c C-j`. There you will find
- `e`: insert last sexp
- `d` insert top level defun (not modified yet to work with `comment` forms intelligently yet
- `r` insert region
- `n` insert ns form