[Fix #2203] Rename cider-interaction.el to cider-eval.el
All of the non-eval functionality was removed, which
marks the end of the process of breaking up the old file.

Down the road we may opt to break down cider-eval into some core
functionality and extra functionality, as many of the eval
commands there are kind of "exotic", but that's not that important
at this point.