Force ns caching for ns lookup related to dynamic indentation
That's another take at
https://github.com/clojure-emacs/clojure-mode/issues/478. The original
solution enabled ns caching globally in `clojure-mode`, but this
created problems for people who'd change a namespace without
noticing they have to invalidate the cached value.

As that's the only real hotspot for `clojure-find-ns` (used internally
in `cider-current-ns`) we can simply let-bind the caching in this
scope and leave evaluations out of this.