Fix incorrect values.
The values for `cider-preferred-build-tool` were quoted, which resulted in the
variable getting the value `(quote lein)`.  This caused an error in
`cider-maybe-intern` when checking the value for stringness.

`cider-allow-jack-in-without-project` had a similarly quoted value, and a safe
property that checked for strings, even though the value should be a symbol.