Minimise the number of requests made during REPL init
New option, cider-repl-init-code, is a list of strings of Clojure code to
evaluate when starting a REPL (with bindings for set!-able vars in place).

Options cider-repl-print-length and cider-repl-print-level are now obsolete.

Set the initial namespace of the REPL using the :ns response slot of an empty
eval message, rather than evaluating (str *ns*). The only other request now made
during init is to evaluate cider-repl-init-code.