Stop cursor moving when initialising the CIDER REPL
Fixes this bug: When cider-repl-pop-to-buffer-on-connect is nil the
cursor jumps when initialising the CIDER REPL. The cursor jumps in
whatever buffer is current.

This commit fixes commit e0aca78b, whose commit message says:

    Make sure REPL window point is at point-max at startup
    Needed when cider-repl-pop-to-buffer-on-connect is 'display-only

This commit applies the new functionality only when the REPL buffer is
being displayed in a window.