[Fix #1840] Add a prototype of find usages
Based on ideas outline here
https://metaredux.com/posts/2019/05/04/discovering-runtime-function-references-in-clojure.html

The code and the UI are a mess, but the feature mostly works. I've
based the UI on apropos, as this seemed the most reasonable given the
fact that we get the references as a list of vars.

There's plenty of room for improvement, but I don't know if I'll find
time for this soon.