Remove cider-latest-middleware-version
There's rarely a case where the latest and the required versions would
be different, so it's simpler to just stick to one version - the
required one.

There was also the problem of specifying a version that's newer than
the required, but still not compatible in some manner. That suppress
the warning, but the problems would still be there.