Ensure fake file paths have a "real" root
CIDER calls `cider-format-connection-params` with the connection
params to make a name for a connection buffer. This calls
`abbreviate-file-name` which checks to see if the filesystem is case
insensitive or not with `file-name-case-insensitive-p`. This is a
function in C that has the following comment in it:

  /* If the file doesn't exist, move up the filesystem tree until we
     reach an existing directory or the root.  */

Introduced in 81d6418e6b7c3e637dccf9c856d9c4b94bd43b97 with the commit
message

Fix file-name-case-insensitive-p on non-existent files

* src/fileio.c (Ffile_name_case_insensitive_p): If the file
doesn't exist, move up the filesystem tree until an existing
directory is found.  Then test that directory for
case-insensitivity.  (Bug#32246)

The important thing here is that when tests are run with Cask we give
it a filename "some/project" and "path/to/dirA/". Neither of these
exists and we go into an infinite loop trying to move the filetree to
an existing file to determine if the filesystem is case sensitive or
not.

Interestingly, this does not happen if I interactively call the
function `(file-name-case-insensitive-p "some/project")` but only when
run under Cask while running the tests.