Bump versions in RTD so that we can rebuild the docs with Python 3
CIDER was running with Python 2 in ReadTheDocs which was causing
downstream build failures when we tried to build the docs without any
pages. After upgrading, mkdoc was still pinned to 0.17.3. My new project
with identical (?) settings in the console was using mkdocs<1.1 as it's
dependency. I couldn't budge the mkdocs dependency in CIDER, even after
wiping the build. Specifying the requirements.txt seemed to be the only
way to override it.