Inherit fringe indicator color from theme
Fringe indicator colors should be easily visible, and complement the current
theme. They should also be green, wherever possible.

The `diff-added` face has a green foreground in almost all themes (all that I
know about) and it should fit the theme closer than an arbitrary green. Inherit
from it where possible, but defer to a hard-coded green if the color is not
easily distinguishable from the background.

The foregound for `diff-added` may be designed to work with a background color.
We could inherit this too, but it would put a lot of visual action into the
fringe. To compromise, we just inherit the foreground.